services/timeseries-writer/requirements.txt
# ClickHouse Writer Service Dependencies

# gRPC
grpcio>=1.62.0  # protobuf 5.x compatibility
grpcio-tools>=1.62.0
protobuf>=5.29.2  # Security update

# Message Queue
pika==1.3.2

# ClickHouse
clickhouse-driver==0.2.6

# Redis (for deleted analysis cache)
redis==5.0.1

# PostgreSQL (for workload sync - Phase 8)
psycopg2-binary==2.9.9

# Logging
structlog==24.1.0

# Configuration
python-dotenv==1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

